+++ DynamicArrays/DynamicArrays/FindSumBetweenNegaative.cpp
@@ -1,5 +1,5 @@
 #include "FirstTasks.h"
-int32_t FindNegative(int32_t* array, int32_t size, bool a)
+int32_t FindNegative(double* array, int32_t size, bool a)
 {
 	if (a == 0)
 	{
@@ -22,11 +22,11 @@ int32_t FindNegative(int32_t* array, int32_t size, bool a)
 		}
 	}
 }
-int32_t SummBetweenNegative(int32_t* array, int32_t size)
+double SummBetweenNegative(double* array, int32_t size)
 {
 	int32_t first_negative{ FindNegative(array, size,0) + 1 };
 	int32_t last_negative{ FindNegative(array, size,1) };
-	int32_t sum{};
+	double sum{};
 	for (int32_t i{first_negative}; i < last_negative; ++i) 
 	{
 		sum += array[i];
